Check if user own block data

+++ src/AppBundle/Controller/ApiController.php
@@ -29,8 +29,10 @@ class ApiController extends Controller
 
         $blockData = $em->getRepository('AppBundle:BlockData')->find($block_id);
 
-        $em->remove($blockData);
-        $em->flush();
+        if ($this->isUserOwnBlock($blockData)) {
+            $em->remove($blockData);
+            $em->flush();
+        }
 
         return new Response();
 
@@ -211,6 +213,16 @@ class ApiController extends Controller
         $em->flush();
 
         return new Response();
+    }
+
+    private function isUserOwnBlock(BlockData $blockData) {
+        $cv = $this->getDoctrine()->getManager()->getRepository('AppBundle:CV')->findOneById($blockData->getCv()->getId());
+
+        if (is_null($cv)) {
+            return false;
+        }
+
+        return $cv->getUser()->getId() === $this->getUser()->getId();
     }
 
 }
